requirements.txt
anyio==4.13.0
certifi==2026.2.25
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
python-dotenv==1.2.2
python-telegram-bot==22.7
